Uses int64_t for trial counts
Increases the maximum number of trials that can be run by using a 64-bit integer to store the trial count.

This prevents integer overflow issues when running simulations for more than 2.1e9 steps.

Fixes bug where assertion for using `cycles_to_complete` with `FlatHistogram` was never executed.

+++ plugin/monte_carlo/include/monte_carlo.h
@@ -7,6 +7,7 @@
 #include <string>
 #include <memory>
 #include <map>
+#include <cstdint>
 
 namespace feasst {
 
@@ -266,7 +267,7 @@ class MonteCarlo {
   virtual void reset_trial_stats();
 
   /// Run a number of trials.
-  virtual void run_num_trials(int num_trials);
+  virtual void run_num_trials(int64_t num_trials);
 
   /// Run until a number of particles is reached.
   virtual void run_until_num_particles(const int num_particles,
